Fixed provenance code duplication
