+++ ui/src/features/reactions/ReactionEntities/entityFormConfiguration/provenance/recordModified.models.ts
@@ -14,63 +14,27 @@
  * limitations under the License.
  */
 import {
-  ReactionFormNodeType,
   type ReactionFormNode,
+  ReactionFormNodeType,
 } from 'features/reactions/ReactionEntities/reactionEntities.types.ts';
-import { wrapInputsWithGrid } from 'common/utils/reactionForm/wrapInputsWithGrid.ts';
+import { createReactionPerson, createUpdatePersonInfoRow } from './reactionPerson.models.tsx';
 
 export const recordModified: Array<ReactionFormNode> = [
-  wrapInputsWithGrid(
-    {
-      type: ReactionFormNodeType.dateTime,
-      name: 'time',
-      wrapperConfig: {
-        label: 'Time',
+  {
+    type: ReactionFormNodeType.wrapper,
+    grid: 2,
+    fields: [
+      {
+        type: ReactionFormNodeType.dateTime,
+        name: 'time',
+        wrapperConfig: {
+          label: 'Time',
+        },
       },
-    },
-    {
-      type: ReactionFormNodeType.value,
-      name: 'person.username',
-      inputType: 'string',
-      wrapperConfig: {
-        label: 'Username',
-      },
-    },
-    {
-      type: ReactionFormNodeType.value,
-      name: 'person.name',
-      inputType: 'string',
-      wrapperConfig: {
-        label: 'Name',
-      },
-    },
-  ),
-  wrapInputsWithGrid(
-    {
-      type: ReactionFormNodeType.value,
-      name: 'person.email',
-      inputType: 'string',
-      wrapperConfig: {
-        label: 'Email',
-      },
-    },
-    {
-      type: ReactionFormNodeType.value,
-      name: 'person.orcid',
-      inputType: 'string',
-      wrapperConfig: {
-        label: 'Name',
-      },
-    },
-    {
-      type: ReactionFormNodeType.value,
-      name: 'person.organization',
-      inputType: 'string',
-      wrapperConfig: {
-        label: 'Organization',
-      },
-    },
-  ),
+    ],
+  },
+  createUpdatePersonInfoRow('person', 'I modified the record'),
+  ...createReactionPerson('person'),
   {
     type: ReactionFormNodeType.value,
     name: 'details',
