+++ ui/src/features/reactions/ReactionEntities/entityFormConfiguration/reactionEntityToForm.ts
@@ -18,7 +18,7 @@ import { reactionNotes } from './reactionNotes.model.ts';
 import { reactionData } from 'features/reactions/ReactionEntities/entityFormConfiguration/data/reactionData.models.tsx';
 import { reactionInputs } from './inputs/reactionInputs.model.tsx';
 import { reactionIdentifiers } from './reactionIdentifiers.ts';
-import { reactionProvenance } from './provenance/reactionProvenance.models.tsx';
+import { reactionProvenance } from './provenance/reactionProvenance.models.ts';
 import { recordModified } from './provenance/recordModified.models.ts';
 import { reactionComponents } from './components/reactionComponents.model.tsx';
 import { reactionComponentsPreparations } from './components/preparations/reactionComponentsPreparations.model.ts';
